Add Support for Sauce Connect & Safari
Safari does not respect configured system proxies for addresses that resolve to localhost.
This prevents Karma from loading in the Safari instance under control during a Sauce Labs test.

This change forces the Travis-CI machine to respond to `karmalocal.dev`.
It also uses this as the hostname of the Karma test server.

To avoid issues with Sauce Connect's SSL Certificate replacement (which is known to
cause Karma grief), this is turned off.